+++ src/StateSmith/Output/Algos/Balanced1/AlgoBalanced1Settings.cs
@@ -19,4 +19,9 @@ public class AlgoBalanced1Settings
 
     public bool generateEventIdToStringFunction = true;
     public bool generateStateIdToStringFunction = true;
+
+    /// <summary>
+    /// https://github.com/StateSmith/StateSmith/issues/141
+    /// </summary>
+    public bool enableDispatchEventResult = false;
 }